Replace ScriptProcessorNode with AudioWorkletNode
The ScriptProcessorNode is deprecated and should be replaced with
AudioWorkletNode.

An AudioWorkletNode can be created with a specific
AudioWorkletProcessor. The AudioWorkletProcessor parses audio data
from the source node and uses a message port to send data back to
the corresponding AudioWorkletNode. Each piece of audio data
is a Float32Array containing 128 elements.