Update UI Interactive Logic
When user has connected to the WebSocket server using the address from the
input element, the connectButton should be disabled.

When user updates the WebSocket address, the connectButton should be
re-enabled. If the user is recording, it should be stopped. If a connection has
been established, it should be closed.

When user types 'Enter' in the input element, it should trigger the 'connect
WebSocket' operation for better UI interaction.